# Use built-in casts from float/double<->word32/word64.