They are available on ghc-4.11 (8.4) and newer
(despite what the comment says, they weren't released in 8.2):
https://github.com/ghc/ghc/blob/1c9dab22f010352884d6f7e5b62251bc4b0cca7c/libraries/base/GHC/Float.hs#L1323

In addition to simplifying our code, this speeds up the float benchmarks.
When rebased onto #302, it provides further speedups:

float-packed(4003B)/encode: 12.29 μs => 7.347 μs
float-packed(4003B)/decode/whnf: 11.13 μs => 4.623 μs
float-unpacked(5000B)/encode 20.5 μs => 17.13 μs
float-unpacked(5000B)/decode/whnf 10.83 μs => 10.90 μs

It's still not great that we're doing the cast per-element, but this
seems like an easy win over something more complicated.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/google/proto-lens/303)
<!-- Reviewable:end -->